Wire up ahjoor-token-whitelist contract and remove dead code
- Add ahjoor-token-whitelist contract (client + tests); escrow,
  payments, and refund now depend on it in place of their local
  no-op TokenWhitelistClient stubs
- Re-enable the previously disabled test_token_whitelist modules
  in escrow, payments, and refund
- Escrow: fix mismatched event name
  (emit_inspection_result_submitted -> emit_inspection_report_submitted),
  drop the unused N-of-M release-approval function
- Refund: drop duplicate DataKey2 enum definition
- Rosca: remove unused events (slot bidding, migration, proxy
  expiry) and unused storage-key enums in savings-goal impl
- Simplify CI: ci.yml runs build+test only (drop coverage-gate job),
  remove audit.yml, .editorconfig, and .cargo/audit.toml

+++ contracts/ahjoor-rosca/src/lib.rs
@@ -10699,7 +10699,7 @@ impl AhjoorContract {
                 for idx in 0..n {
                     if idx == i {
                         new_items.push_back(items.get(j).unwrap());
-                    } else if idx as u32 == j {
+                    } else if idx == j {
                         new_items.push_back(items.get(i).unwrap());
                     } else {
                         new_items.push_back(items.get(idx).unwrap());
